+++ src/redux/sagas/devices.saga.js
@@ -36,12 +36,25 @@ function* fetchDevicesByUserId() {
   }
 }
 
+// worker Saga: will be fired on "ADD_DEVICE" actions
+function* addDevice(action) {
+  try {
+    console.log("IN SAGA: ", action.payload)
+    // passes the username and password from the payload to the server
+    yield axios.post('/api/devices', action.payload);
 
+    // automatically log a user in after registration
+    // yield put({ type: 'DEVICE', payload: action.payload });
+
+  } catch (error) {
+    console.log('Error with adding device:', error);
+  }
+}
 
 function* DevicesSaga() {
   yield takeLatest("FETCH_DEVICES", fetchDevices); // not being used. get all devices for all users
-  yield takeLatest("FETCH_DEVICES_BY_USER_ID", fetchDevicesByUserId)
-
+  yield takeLatest("FETCH_DEVICES_BY_USER_ID", fetchDevicesByUserId);
+  yield takeLatest('ADD_DEVICE', addDevice);
 }
 
 export default DevicesSaga;
